[AT-Factory-Tool] Support configuration file.
The tool now read configurations from a configuration file. The
configuration change would also be saved to the file so that it
would not be lost if you close the tool. For example, if you
change low key alert threshold, the tool would remember it in the
configuration file.
The configuration includes:
ATFA_REBOOT_TIMEOUT: The reboot timeout for an ATFA device. User should
not change.
ATFT_VERSION: The version for the Android Things Factory Tool. User
should not change.
COMPATIBLE_ATFA_VERSION: The compatible ATFA version for this tool.
User should not change.
DEFAULT_KEY_THRESHOLD: The default low key alert threshold. If the
remaining key is lower than this value, an alert would be given. This
setting would be overwritten by the in-tool settings.
LANGUAGE: The language for the tool. Possible values are 'cn' (Chinese)
and 'eng' (English).
LOG_DIR: The log directory.
LOG_FILE_NUMBER: The limitation for log file number.
LOG_SIZE: The limitation for log size.
PRODUCT_ATTRIBUTE_FILE_EXTENSION: The file extension for product
attribute file, user should not change.
REBOOT_TIMEOUT: The reboot timeout for target device. If the target
device takes a long time to reboot into fastboot mode, user should
increase this value.

Test: Manual test and unit tests.
Bug: b/68662667, b/65837743
Change-Id: I09b0160ca777baedd9da3edbf192d51fa5887fc8