[ATFT] Fix a issue when failed operation does not stop state transition.
In _HandleStateTransition function, the state transition processing
loop should end if one of the operation failed, however, the target
variable might change after each operation. If the first operation
fails, the target is not updated until the while not failed check
passes. During that check, we are checking the old target device not
the new target device. As a result, failed operation would not be
detected and we would continue the rest of the operations.

Test: Unit tests. Add test testHandleStateTransitionTargetDeviceChange
and testHandleStateTransitionRebootTimeout.
Bug: None
Change-Id: I8be872f3920e573e93e9594a77d131e8f68cbb3a