[ATFT] Add SINGLE_DEVICE mode.
Add a single device mode when no usb location mapping is present.
In this mode, we assume the user would only use one target device.
If multiple target devices is detected in SINGLE_DEVICE mode, user
has to either unplug one device or map usb locations, which would
turn the mode to MULTI_DEVICE mode, which was the previous default
mode. In this mode, each used USB location has to be mapped to one
UI slot.

This CL also adds the skip product button to allow skip choosing
product file, however, some operations might not be avilable if
user skip choosing the product file. The automatic mode is also
not available if skipped.

This CL also fixes a bug when you could do operation to a rebooting
target device, a rebooting target device should not be able to
do any operations.

Tests: manual tests on Linux and Windows.
Bug: b/118001288
Change-Id: Ic42bb2e7f153f097edc53fa0735fcdf6ef708b53