[ATFT] Add 'Start Operation' button.
Add a button to control automatic operation in operator mode. During
development, we find that start auto provisioning when the tool is
opened sometimes cause unintended behavior, for example, we do not
want to lock the avb but the tool lock it before we could enter
supervisor mode. So I add a button to control that. User now has
to click 'Start Operation' button in operator mode to actually start
the auto provisioning.

Test: Manual test.
Bug: b/118001288
Change-Id: I86297eb20357145ee0cc1d2ae229059a6a443a4a