[ATFT] Make file backward compatible for v2.0.
Add logic to check config file version, if config file version is
v2.0 and the software version is v3.0, we would upgrade the config
file to v3.0. User could upgrade from v2.0 to v3.0 and keep all the
configurations including mapping status by copying the config file.

Test: 1. Verify config version upgrades from v2.0 to v3.0.
2. Verify that config version stays the same even if it is invalid.
3. Verify that config files with wrong version (not v2.0 or v3.0)
would be rejected and the config file version would not be touched.
Bug: None
Change-Id: Id1a4c82b35a31eaaccd88b2ec002f41833b6e1ad