[ATFT] Add option to skip rebooting after fusing vboot key.
For some SoC 'oem fuse at-bootloader-vbootkey' would trigger a
reboot, in this case, another reboot is not necessary after this
command. In order to save overal time in the flow, we now add option
to skip rebooting after fusing vboot key. To configure this, set
'SKIP_REBOOT' to True in the configuration file.

Tests: Manual test on test target device that would reboot after
fusing command.
Bug: 122065121